requirements.txt
django
djangorestframework
django-cors-headers
PyJWT
mysqlclient
django-cors-headers
mvola
python-dotenv
djangorestframework-simplejwt

